# tests: Add perf test for shard_token_bucket