@xemul 2023-07-29
upd:
- added capped-release test that finally shows a real problem