@xemul 2023-07-31
upd:
- removed some unused stats
- less hard-coded constants
- relaxed hogs not to burn CPU throughout the whole task-quota
- improved releaser to work like real reap_kernel_completion -- it releases tokens proportional to real pause duration